pyproject.toml
[project]
name = "carechat-planer"
version = "0.1.0"
description = ""
authors = [
    {name = "hoppiece@gmail.com"}
]
license = {text = "GPL-v3"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "line-bot-sdk (>=3.16.0,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "logzero (>=1.7.0,<2.0.0)"
]
[tool.poetry]
packages = [
    {include = "planer_bot", from = "src"},
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.6"
mypy = "^1.15.0"
pytest = "^8.3.4"

